requirements.txt
opencv-python == 4.12.0.88
numpy == 2.3.1
customtkinter == 5.2.2
darkdetect == 0.8.0
packaging == 25.0
Pillow == 11.3.0